[#1] refactoring
